requirements.txt
# Core
streamlit>=1.31.0
python-dotenv>=1.0.0

# LangGraph & LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20

# Databases
pinecone-client>=3.0.0
neo4j>=5.15.0

# Embeddings
google-generativeai>=0.3.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.0
pydantic>=2.5.0

# Data Processing
aiohttp>=3.9.0
